requirements.txt
google-cloud-storage>=2.17.0
google-cloud-bigquery>=3.20.0
pylint>=3.2.3
pytest>=8.2.2
pytest-cov>=5.0.0
ipython>=8.25.0
black>=24.4.2
google>=3.0.0
typing>=3.7.4.3
httpx>=0.27.0
python-dotenv
openlineage-airflow
apache-airflow-providers-ssh
apache-airflow==2.9.2
apache-airflow-providers-google==10.19.0
apache-airflow-providers-openlineage==1.8.0
apache-airflow-providers-common-sql>=1.6.0
openlineage-integration-common>=0.28.0
openlineage-python>=0.28.0
attrs>=22.2
setuptools==66.1.1